refactor(search): add SearchEntry type to shared/
- shared by various search-related modules as well as frontend in the future
