+++ shared/src/types/api/search.ts
